:recycle: Refactor common fixtures
