+++ tests/test_api.py
@@ -13,55 +13,6 @@ URLS = [
 ]
 
 
-@pytest.yield_fixture
-def releases(client):
-    rel = {}
-    for i in range(5):
-        r = client.post(BASE_URL+'/releases',
-                        {'name': 'TEST', 'studies': 'SD_{0:08d}'.format(i)})
-        rel[r.json()['kf_id']] = r.json()
-    return rel
-
-
-@pytest.yield_fixture
-def task_services(client):
-    ts = {}
-    with patch('coordinator.api.validators.requests') as mock_requests:
-        mock_resp = Mock()
-        mock_resp.content = str.encode('{"name": "test"}')
-        mock_resp.status_code = 200
-        mock_requests.get.return_value = mock_resp
-
-        for i in range(10):
-            r = client.post(BASE_URL+'/task-services',
-                            {'name': 'TASK SERVICE {}'.format(i),
-                             'url': 'http://localhost',
-                             'description': 'test'})
-            ts[r.json()['kf_id']] = r.json()
-    return ts
-
-
-@pytest.yield_fixture
-def tasks(client, releases, task_services):
-    ta = {}
-    for i in range(50):
-        rel = releases[random.choice(list(releases.keys()))]['kf_id']
-        ts = task_services[random.choice(list(task_services.keys()))]['kf_id']
-        r = client.post(BASE_URL+'/tasks',
-                        {'name': 'TASK {}'.format(i),
-                         'release': BASE_URL+'/releases/'+rel,
-                         'task_service': BASE_URL+'/task-services/'+ts})
-        ta[r.json()['kf_id']] = r.json()
-    return ta
-
-
-@pytest.yield_fixture
-def fakes(releases, task_services, tasks):
-    return {'releases': releases,
-            'task-services': task_services,
-            'tasks': tasks}
-
-
 @pytest.mark.parametrize('endpoint', URLS)
 @pytest.mark.parametrize('method', ['get', 'post', 'patch', 'put'])
 def test_response_codes_200(client, db, endpoint, method):
